Resume the pipeline when a new zone is loaded
If a zone's pipeline is soft-halted, and a new version of the zone is
loaded, the pipeline doesn't resume.  Part of the reason for this was
to stop the zone signer from signing an unsigned zone that hasn't been
approved.  A secondary issue is that a 'SignZone' command (e.g. issued
autonomously by the key manager), while the pipeline is halted, causes
the zone signer to hard-halt.  This accidentally turns soft-halts into
hard-halts.

Now, a new zonetree for ready-to-sign zones has been added.  Unsigned
zones are added to this tree when they are approved.  This separates
the pipeline mode from the unsigned zone approval state; the zone signer
can be run at any time, and it will only touch approved zones.  The zone
signer has also been adjusted to leave the pipeline mode alone if it is
already halted.