Wait for other units before taking sockets in "PS".
Introduce a ready channel for units to notify the manager that they are
ready, and use this to wait for other units to become ready before
spawning the "PS" unit. This allows "PS" to reliably give other units a
chance to take systemd provided sockets for themselves before "PS" takes
whatever is left over.