Review feedback: Remove the accept-systemd-sockets config setting.
Always accept them if they are provided, as it is very unlikely that the
config and systemd settings are managed by different people or that a
sys admin with access to systemd settings doesn't also have access to
the Cascade config file, or that an operator configures socket
activation for Cascade by accident. Also, having acceptance needing to
be enabled adds a confusing step for operators who want to enable
passing of sockets from systemd to Cascade.