Ensure that second attempt to sign zone A while zone A is being signed doesn't lock up.
Attempts to write to signing progress metrics would block if a second
signing attempt for the same zone was enqueued, as enqueue() locked the
whole signing progress metrics state until it could join the queue,
which for the same zone is only possible once current signing of the
zone finishes -- i.e. never!