Neither skip re-approval nor skip zone promotion through the pipeline.
In the case of a zone that failed review, on zone reload we should:

a. Re-review it, because the content may have changed.
b. Allow it to progress to the next stage in the pipeline (if review
passes).

Prior to this change it was neither re-reviewed nor progressed, new zone
content triggering a re-review was just ignored.