Serve fake RRSIG RRSET with corrected TTL. (#214)
To work around the zone tree storing RRSIGs as an RRSET with a single TTL which is incorrect for RRSIGs.

Note: Switches from Zone to LightWeightZone so that the zone walk behaviour can be overridden, which means we lose the more proper DNS response generation, but gain reduced memory consumption (but we already use LightWeightZone for XFR received zones anyway so this is not entirely new).